codesentinel/requirements.txt
langgraph==1.0.10
langchain-openai==1.1.10
langchain-groq>=0.2.0

PyGithub==2.8.1
fastapi>=0.115.0,<0.120.0
uvicorn[standard]>=0.32.0,<0.35.0
httpx>=0.27.0,<0.29.0
SQLAlchemy>=2.0.36,<3.0.0
greenlet>=3.0.0
aiosqlite>=0.20.0,<0.21.0
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
tenacity>=8.2.0,<10.0.0

pytest>=8.3.0,<9.0.0
pytest-asyncio>=0.24.0,<0.26.0
